requirements.txt

langchain
langchain-openai
langchain-anthropic
langgraph
python-telegram-bot
anthropic
openai
elevenlabs
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pinecone-client
pyairtable
pydub
SpeechRecognition
gspread
python-dotenv
requests
